requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
attrs==24.2.0
azure-cognitiveservices-speech==1.40.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
moviepy==1.0.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.1.1
openai==1.50.2
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipenv==2024.1.0
pipreqs==0.5.0
platformdirs==4.3.6
proglog==0.1.10
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.6
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
boto3==1.35.30
